Rewrite parser with a closer to grammar logic
Previous version did not capture all possible productions that the grammar
allows. This new version follows grammar productions much more closely and
generates an abstract syntax tree that is also closer to these grammar
productions.

The AST may need to be simplified a bit (not clear that the `GroupEntry`
and `GroupChoice` levels are that useful in particular), but make it possible
to express valid CDDL files more easily.