Another rewrite to avoid duplication in AST
The AST was composed of parent nodes with a list of children. That was
fantastic to store tokens with whitespaces and comments. But this meant that
content was duplicated across all nodes as nodes also stored a more semantic
representation of the underlying content.

This update drops the parent/children relationship where ever possible and
rather uses the (completed) semantic representation as the basis for the
serialization. This allows us to consider making updates to some node in the
tree, and still expect these updates to be part of the serialization
afterwards.

This is done by making most nodes inherit from a new abstract class
`TokenNode` to store whitespace and comments that may arise before the node,
and a possible separator token that may be used after it to separate it from
the next item in an underlying list.

Such nodes may also contain an open token and a close token.

The update also completes the list of known operator names with those defined
in RFC9165: https://datatracker.ietf.org/doc/html/rfc9165

Notes:
- The `Memberkey` and `Occurrence` nodes remain clunky for now and continue
to store a list of Tokens on top of their semantic representation.
- Some node values are now represented as Token whereas processing would be
easier if they were captured as strings or numbers directly.