Add initial logic to convert GroupEntry to Type
The logic is clunky and incomplete. I'm not clear I understand the different
reasons why a rule is to be regarded as a type rule or as a group rule.

Also, an intrinsic limitation is that the code needs a Group to represent the
construct "(" S type S ")". That should rather be represented as a Type too.