Fix parsing of identifiers, text and byte strings
The parser was essentially chewing any character between wrapping quotes
both for text and byte strings. It now correctly follows the grammar and
reports an error if some illegal character is used.

Similarly, identifiers have specific rules that were not checked (for example,
start character cannot be a digit, and identifier cannot end with "-" or ".")