Add custom __repr__ methods for the AST tree
Something changed in Python 3.10 in prettyprint that breaks the testing logic.
Since that logic was painful in that the produced serialization was barely
readable, this update implements a custom nested string representation for the
tree, for use in tests and debugging.

Note that, for each node that can have open/close tokens, the representation
starts with those tokens. Similarly, if a node ends with a separator, that
separator appears first. The rest is more or less as expected.