Improve syntax validation error messages
The position at which parsing fails is usually at the end of the token that
follows the token that is actually problematic. Pending further improvements to
error reporting, the exact position turns out to be more confusing than
anything else. The update makes the parser stick to reporting just the line
number where a problem occurred.

The CDDL extract is no longer reported either. Instead, messages were
re-written to be more human-friendly and to include the actual problematic
construct (stripped of whitespaces) right away.

The messages also use markdown backticks for code extracts to ease integration
in systems that might want to render the error.