Fix __repr__ method for Memberkey, improve it for Typename
The representation of a Memberkey did not output the key type name, only
the possible caret and colon that follows it.

The representation of Typename was optimized for constructs without whitespaces
but the representation was barely readable when whitespaces are actually
present. The __repr__ method now always reports the Typename name on its own
line.

Test snapshots were re-generated accordingly.